Move grading test to events suite
pylint correctly surmised that `test_grade` is an event test, so move it
over.
